Add view of full statistics

+++ src/main/java/org/example/dto/Type.java
@@ -1,5 +1,5 @@
 package org.example.dto;
 
 public enum Type {
-    LONG, DOUBLE, STRING
+    INTEGER, FLOAT, STRING
 }
